refactor: split workspace.rs god-file into modules
